Don't assign Firefox a default workspace
Doesn't make sense to do this, actually. It's nice to have it assume a
default workspace on start, but then the assignment includes at
runtime. So opening a link in a new tab or window sends it to ws1… not
always what I want. Also, the behavior I want is to start-up by default
in ws1. This already happens since ws1 is the default workspace on
login.