[ansible] Transition from monolithic playbook to roles
This commit is huge and changes many files, but really, it only
rearranges everything and breaks them into reuseable, modular pieces.
This way, it's possible to repeat some app-level configs between
different operating systems types. It also makes it easy to "opt in" to
some roles and exclude others. The possibilities are endless!

Anyways, this is a big change, but it looks scarier than it is.

Closes #6.