Misc. role cleanup and improvement of user experience (#17)
* [sys/centos,rhel] modularize tasks into categorized groupings

This commit intends to reduce complexity in the main task file. Related
tasks are grouped together in separate, named files.

* [sys/fedora] Update to loop syntax, minor pedantic fixes

* [task] Consolidate taskrc files to single file with jinja2 boolean logic

* [vagrant] Upgrade to F29 image, set compatibility_mode

* [system/base] import global variables explicitly

* [system/*, thunderbird] use hostname lookup to determine config path

Eliminate all need for a ~/.jwf-identifier file. It always returns a
changed value in a playbook run (not helpful) and it's more reliably
replicated by setting hostnames to serve as differentiators.