system/fedora: Install planner by default
The `planner` package is an open source project management utility
available as part of the GNOME Desktop. I started to use it for ISTE-500
Senior Development I at the Rochester Institute of Technology this
semester as project manager for our project. It is useful so far and
fairly easy to use. The keyboard shortcuts are essential to faster data
entry (e.g. importing tasks from another document).

Anyways, I think this one is a keeper for now. :+1: