system: Create packaging-tools role for RPM packaging
This commit creates a new `packaging-tools` role. The role is intended
to include my preferred configuration for RPM packaging. Currently, it
assumes to be used only on a Fedora Workstation, but this may change in
the future.

It reduces some of the packaging-specific steps from the base
system/fedora-workstation role.