:cd: ripright: Add container-based workaround with F29
This commit adds two new Dockerfiles with instructions to build Fedora
29 containers with the `ripright` package installed. The `ripright`
package was orphaned and subsequently abandoned in Fedora starting in
Fedora 30. The newer gcc version was incompatible with ripright. The two
Dockerfiles represent a clever workaround by creating a F29 container,
mounting a physical CD/DVD drive into the container, and ripping the
contents of a CD from a container to the host system.

Neat-o!

Right now, this might not work fully as expected because of on-going
development in rootless containers. The intention is to run this without
root or any superuser permissions required. Since Podman's rootless
containers use Linux kernel user namespaces, this should be possible at
some point.

Signed-off-by: Justin W. Flory <git@jwf.io>