group_vars: Set ansible_python_interpreter to python3
This change is required for the `package` module to work on a Fedora 30
system using Ansible 2.8. It is required because the `package` module
tries to install `python2-dnf` which was retired in Fedora 30. In order
for `package` modules to work from a F30 base system, this config change
is required.

Shout-out to @neingeist for sharing this fix in ansible/ansible#54855:

    https://github.com/ansible/ansible/issues/54855