Add Pipfile for virtualizing ansible binary
On Fedora Workstation, I always have the latest version of Ansible but
this is not always the case when I am using another distribution like
Red Hat Enterprise Linux/CentOS. Using a Pipfile allows me to use the
latest upstream version of Ansible regardless of platform.

I hit this issue originally because some of the syntax I use in my roles
(i.e. `loop`) is specific to newer versions of Ansible.

Signed-off-by: Justin W. Flory <git@jwf.io>