inventory: Remove deprecated characters from group name
This commit resolves a deprecation warning that comes from using invalid
characters in a group name in the inventory. Hyphens are a no-no, it
seems. So, I just changed it to an underscore.

Solves this warning specifically:

```
[DEPRECATION WARNING]: The TRANSFORM_INVALID_GROUP_CHARS settings is set
to allow bad characters in group names by default, this will change, but
still be user configurable on deprecation. This feature will be removed
in version 2.10.

[WARNING]: Invalid characters were found in group names but not
replaced, use -vvvv to see details
```

Signed-off-by: Justin W. Flory <git@jwf.io>