roles: Add new Slack role
This commit creates a new Slack role and adds it to the Fedora
Workstation playbook. This change was motivated by a mismatched Slack
config that broke GPG signing on the Slack package on my laptop for
ages. So, I figured it makes sense to start tracking this in swiss-army
too.

It's a very simple role. It copies a static, fixed file into the correct
place and installs the `slack` package.

Signed-off-by: Justin W. Flory <git@jwf.io>