inventory/vars: Add target_group variable
This commit adds `target_group` as a new variable in the global
environment. This is intended to correspond with environments where I
may want these two values to be different, like they were at one of my
previous employers with a RHEL 7 environment.

This does not change all places where I reference `group`, but it puts
the variable in existence so I can start writing new roles correctly,
and migrate the others later.

Signed-off-by: Justin W. Flory <git@jwf.io>