slack: Use updated repo file
For some reason, Slack ships the repo file as committed previously. But
whenever you actually interact with the repo from the command line with
DNF, it overwrites the path with these values. It also prompts me to
re-accept the GPG signing keys. So I hope this makes me less annoyed.

Signed-off-by: Justin W. Flory <git@jwf.io>