:policewoman: bash: Semantics with auto-completion scripts
This is a semantic policing. So, the task.sh file is a bash-completion.
So the file extension could more clearly indicate that. I followed the
precedent set by the Poetry project for using this file extension. I
added a step in my base workstation role to ensure the old file is
removed. Because… OCD. :eyes: Don't judge me pls.

Signed-off-by: Justin W. Flory <git@jwf.io>