:shield: bash: Harden file permissions
This commit changes some of the file permissions on my Bash shell files.
Specifically, it sets an SELinux user context on all the files since the
system, Ansible, is modifying the files. It also makes some files read
only that are managed in Ansible, to discourage editing them directly.
Edits should be managed and pushed out in config management always.

Signed-off-by: Justin W. Flory <git@jwf.io>