:repeat: bash: Use system bash_completions.d directory (#35)
This commit changes behavior for my Bash auto-completion files. The
directory at `~/.bash_completions.d` is not a default path for Bash
auto-completions. Adding the files into `/etc/bash_completions.d` caused
an instant change in auto-completion in a new shell. Instead of trying
to hack this together, this commit uses the system directory for loading
these files into the shell.

It also purges the old directory from the home directory and cleans up
the old `source` commands in my `bashrc` file.

Signed-off-by: Justin W. Flory <git@jwf.io>