:card_file_box: Archive Taskwarrior role
I no longer use Taskwarrior as a part of my daily workstation. This
archives the role and puts it away somewhere where it will not get
collected and ran as part of normal playbooks.

Additionally, I moved the "purge unused files" task out of the
apps/base-workstation role and into the system/base role, since this is
a universal task to apply in all environments. I also purged the `task`
package from all systems too.

Signed-off-by: Justin W. Flory <git@jwf.io>