:lock: bash: Narrow down setype context, swap target_group in
I decided to change the `setype` to match the mappings of files and
directories in `~/.config/`. Since these files are at the root of the
home directory, they are not labeled as such, but there is no reason why
they shouldn't be.

Also, this swaps one more `group` from `target_user` to `target_group`.
Huzzah for gradually resolving technical debt!

Signed-off-by: Justin W. Flory <git@jwf.io>