:wastebasket: scripts: Discard ripright scripts/container
Unfortunately, `ripright` is a pain to use on Fedora. I discovered
`whipper`, which is a cool Python3 project that seems to meet the use
case I need for ripping audio CDs. It includes FLAC support and uses the
MusicBrainz database for metadata lookups.

This commit is clearing the way for a new script or way of ripping my
audio CDs.

Signed-off-by: Justin W. Flory <git@jwf.io>