:secret: protonvpn: Add passwordless sudoers rule
This commit adds a file to be dropped in at /etc/sudoers.d/10_protonvpn
for passwordless `sudo` for any users belonging to the `wheel` group. In
simple terms, this commit lets me run `sudo protonvpn` without typing my
password. Because I do run this command all the time. :scream:

Thanks to the manpages:

    https://www.sudo.ws/man/1.8.15/sudoers.man.html

Signed-off-by: Justin W. Flory <git@jwf.io>