:package: system/fedora: Add shotwell to base package set (#64)
So, in an unexpected turn of events, it looks like GNOME Shotwell is
still being actively developed, with some support from the Software
Freedom Conservancy. A while ago, I remember Shotwell was going to be
dropped from Fedora because a new tool was going to replace it. Which
made me sad because Shotwell is the minimal tool I always liked for
making quick photo edits.

But alas, it is being actively maintained in upstream GNOME and it is
currently packaged for Fedora. So I'm going to formally adopt it into my
base package set. :tada:

Signed-off-by: Justin W. Flory <git@jwf.io>