:package: system/{base,fedora}: Swap cmus with lollypop
I gave cmus a shot and it was cool for a while, but Lollypop has really
won me over. It's such a beautiful music player. cmus is cool for a CLI-
based music player, but you lose out on the richness of album artwork
and artist artwork. And y'know… that's really important.

So, bye cmus. It was nice. Hello lollypop.

Signed-off-by: Justin W. Flory <git@jwf.io>