:card_file_box: playbooks: Refactor vars from role to playbooks (#75)
Previously, there was a set of global variables I defined in
`inventory/group_vars/all.yml`. In my `system/base` role, I had a task
that explicitly imported those variables into the playbook run. This was
used across all of my roles.

As I found out recently, when there was a divergence in those variables,
either for a specific host or a type of host, managing the difference
was difficult. Overriding the variables was only possible by doing so on
the command line. Since that is easy to forget, I knew there must be a
better way.

So now, all of these variables are defined uniquely in each playbook.
Not sure if this is _The Best Possible Way_, but it is certainly better
than it was before. Keeping it this way for now!

Yay paying off personal technical debt.

Signed-off-by: Justin W. Flory <git@jwf.io>