:building_construction: Add D.I.Y. Asciidoc docs publishing pipeline (#90)
* :foggy: gitignore: Expand coverage for RubyGem automation

Exploring some new tools to build my documentation into something more
practical.

Signed-off-by: Justin W. Flory <git@jwf.io>

* :gem: Add asciidoctor/html-proofer for planned docs-building pipeline

First step of integrating into some sort of GitHub Pages documentation
publishing. I also added html-proofer for linting generated HTML docs.

Signed-off-by: Justin W. Flory <git@jwf.io>

* :trollface: Add docs build script

This is going to be really hacky, but it might also be neat too. This is
a bit of a gross script because I am using `sed` to filter output. A
proper programming language would make more sense for this.

Signed-off-by: Justin W. Flory <git@jwf.io>

* :alembic: ci: Add job for HTML docs site generation

This is increasingly becoming hackier and hackier.

Signed-off-by: Justin W. Flory <git@jwf.io>

* :arrow_right: ci: Add deploy step for HTML pages to GitHub Pages

Now, the final step that sowes it all together.

Signed-off-by: Justin W. Flory <git@jwf.io>