:memo: how-to-use: Add callout to generate new repo from this template
Since I enabled this feature, other GitHub users can one-click fork this
repo, but instead use it as a template. Instead of inheriting the
history of `git` commits, it creates a new repo where the state of
`HEAD` in the template repository is the first commit of the new repo.

This "forking" model makes more practical sense in how someone might
want to use this repo as a downstream derivative work.

Signed-off-by: Justin W. Flory <git@jwf.io>