:tada: system/fedora: Do not ignore errors from Docker service anymore
Huzzah! Docker now supports cgroups v2. A recent package update to
`moby-engine` gets this working and implemented on Fedora 31. I was able
to get the Docker daemon to start successfully on my F31 Workstation.

So, going to pay off this tech debt now by allowing a failure to stop
the playbook run, as it rightfully should.

Signed-off-by: Justin W. Flory <git@jwf.io>