:lock: ssh: Refactor custom changes as drop-in configuration (#95)
This commit reworks my SSH role to use a drop-in template with only the
changes I need. This drop-in config template is added to
`/etc/ssh/sshd_config.d/` where it is automatically picked up by the
system. This reduces conflicts with distribution defaults that are
shipped by package maintainers to `/etc/ssh/sshd_config`.

So, you learn a few things from `rpmconf -a`.

Signed-off-by: Justin W. Flory <git@jwf.io>