:package: system/flatpak: Initial commit
This commit adds a new Flatpak role to swiss-army. This Flatpak role
takes over some of the more complex software installation in the
fedora-workstation role. Ultimately, the goal in this change is better
unification for how software that is not very Fedora-friendly to be
managed more easily.

The `system/fedora-workstation` role also gains a dependency on the
`system/flatpak` role. This ensures the Flatpak role is run before the
system role is applied. It's a more declarative and handy way to link
roles together, instead of making my playbooks longer and longer.

Add galaxy_info metadata to fix Ansible Lint error

Signed-off-by: Justin W. Flory <git@jwf.io>