:night_with_stars: Fix bugs on a new install
I reinstalled Fedora 32 on my laptop this time. This marks the first
major departure from a hard drive partition I have more or less carried
since Fedora 20. Long live and prosper!

Here are the commits I squashed together:

---

:beetle: pass: Clone git repo as target user to avoid SSH mishap

By default, the root user's SSH key is used to connect to GitHub.
Normally this is fine, but in this case, any old SSH key will not do. It
needs to be one associated to my account since it is a private repo.
This is a band-aid fix for fresher installs.

---

:beetle: fedora-workstation: Create /root/.ssh/ directory

On a fresh install, this directory will not exist by default. Needs to
be initialized in order for an SSH key to be generated. It would be nice
if this were taken care of the SSH module, but oh well.

---

:new: playbooks: Add host-specific playbook for fossbook

This commit creates a new playbook specific for my laptop, made in the
image of `playbooks/fedora-workstation.yml`. The only difference is in
the variables, where I updated my username to a new shorthand username.

---

:new: scripts: Add one-off network scripts migration playbook

A short and simple playbook to unpack a /etc/sysconfig/network-scripts/
backup on a new host. It takes care of file permissions and makes sure
things are permissioned correctly.

A nice companion for this playbook would be a save version, to create a
one-off backup of network scripts on a host.

---

Signed-off-by: Justin W. Flory <git@jwf.io>