:couple: playbooks: Drop host-specific playbook, update Fedora playbook
This commit drops the `fossbook.yml` playbook. It was identical to
`fedora-workstation.yml` except for the variables in the header. Now
that I reinstalled Fedora Workstation on my desktop, the config info is
unified on all my "live" Fedora installs.

a.k.a. back to the way it used to be

Signed-off-by: Justin W. Flory <git@jwf.io>