:peace_symbol: centos: Bring harmony and simplicity, no more duplicity
All of my CentOS environments are now running CentOS 8. Instead of
making a playbook for each individual host (with almost identical
configurations), the end-user is encouraged to use the `--limit` flag of
`ansible-playbook` to limit by a desired hostname.

This makes little to no functional changes, but it reduces overhead by
simplifying how much source must be maintained for a group of
environments that are already (nearly) identical.

Signed-off-by: Justin W. Flory <git@jwf.io>